+++ packages/presets/graphql-modules/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @graphql-codegen/graphql-modules-preset
 
+## 5.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`8258f1f`](https://github.com/dotansimha/graphql-code-generator/commit/8258f1f6012c106d02ef28bca9ec424f70c4aa26)]:
+  - @graphql-codegen/visitor-plugin-common@6.1.0
+
+## 5.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`accdab6`](https://github.com/dotansimha/graphql-code-generator/commit/accdab69106605241933e9d66d64dc7077656f30)]:
+  - @graphql-codegen/visitor-plugin-common@6.0.1
+
+## 5.0.1
+
+### Patch Changes
+
+- [#10447](https://github.com/dotansimha/graphql-code-generator/pull/10447) [`5dad86e`](https://github.com/dotansimha/graphql-code-generator/commit/5dad86e4753369e288fba5821722507d88afb3dc) Thanks [@eddeee888](https://github.com/eddeee888)! - Fix \_\_isTypeOf wrongly picked on objects that are not implementing types or union members
+
 ## 5.0.0
 
 ### Major Changes